+++ src/Applications/use_case/_test/DeleteUserUseCase.test.js
@@ -1,4 +1,5 @@
 const UserRepository = require('../../../Domains/users/UserRepository');
+const AuthenticationTokenManager = require('../../security/AuthenticationTokenManager');
 const DeleteUserUseCase = require('../DeleteUserUseCase');
 
 describe('DeleteUserUseCase', () => {
@@ -8,21 +9,47 @@ describe('DeleteUserUseCase', () => {
       id: 'user-123',
     };
 
+    const useCaseHeader = {
+      authorization: 'Bearer accessToken',
+    };
+
+    const admin = { id: 'user-1' };
+
+    const expectedAccessToken = 'accessToken';
+
     const mockUserRepository = new UserRepository();
+    const mockAuthenticationTokenManager = new AuthenticationTokenManager();
+
     mockUserRepository.checkUserIsExist = jest.fn()
       .mockImplementation(() => Promise.resolve());
     mockUserRepository.deleteUserById = jest.fn()
       .mockImplementation(() => Promise.resolve());
+    mockUserRepository.verifyAdmin = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+
+    mockAuthenticationTokenManager.getTokenFromHeader = jest.fn()
+      .mockImplementation(() => Promise.resolve('accessToken'));
+    mockAuthenticationTokenManager.verifyAccessToken = jest.fn()
+      .mockImplementation(() => Promise.resolve());
+    mockAuthenticationTokenManager.decodePayload = jest.fn()
+      .mockImplementation(() => Promise.resolve(admin));
 
     // Action
     const deleteUserUseCase = new DeleteUserUseCase({
       userRepository: mockUserRepository,
+      authenticationTokenManager: mockAuthenticationTokenManager,
     });
 
-    await deleteUserUseCase.execute(useCaseParams);
+    await deleteUserUseCase.execute(useCaseParams, useCaseHeader);
 
     // Assert
+    expect(mockAuthenticationTokenManager.getTokenFromHeader)
+      .toBeCalledWith(useCaseHeader.authorization);
+    expect(mockAuthenticationTokenManager.verifyAccessToken).toBeCalledWith(expectedAccessToken);
+    expect(mockAuthenticationTokenManager.decodePayload).toBeCalledWith(expectedAccessToken);
+
     expect(mockUserRepository.checkUserIsExist).toHaveBeenCalledWith(useCaseParams.id);
+    expect(mockUserRepository.verifyAdmin).toHaveBeenCalledWith(admin.id);
     expect(mockUserRepository.deleteUserById).toHaveBeenCalledWith(useCaseParams.id);
   });
 });
